+++ drivers/net/mlx5/mlx5_flow.c
@@ -31,6 +31,7 @@
 #include "mlx5_defs.h"
 #include "mlx5_flow.h"
 #include "mlx5_glue.h"
+#include "mlx5_devx_cmds.h"
 #include "mlx5_prm.h"
 #include "mlx5_rxtx.h"
