+++ drivers/net/mlx5/mlx5_rxtx.c
@@ -29,6 +29,7 @@
 #include <rte_flow.h>
 
 #include "mlx5.h"
+#include "mlx5_devx_cmds.h"
 #include "mlx5_utils.h"
 #include "mlx5_rxtx.h"
 #include "mlx5_autoconf.h"
